// src/shared/api/model/templateItemDto.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * 프로필 카드 공유 서비스 API
 * 프로필 카드 공유 서비스 백엔드
 * OpenAPI spec version: 0.1.4
 */

export interface TemplateItemDto {
  /**
   * 항목 이름
   * @minLength 1
   * @maxLength 255
   */
  label: string;
  /**
   * 항목 설명
   * @maxLength 2000
   */
  description: string;
  /**
   * 항목 타입 (0: SHORT_TEXT, 1: LONG_TEXT, 2: LINK, 3: NUMBER)
   * @minimum 0
   * @maximum 3
   */
  type: number;
}